+++ cipher/CHANGELOG.md
@@ -8,12 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.5.0 (UNRELEASED)
 ### Changed
 - Edition changed to 2024 and MSRV bumped to 1.85 ([#1759])
+- Re-export of `crypto-common` moved to `cipher::common` ([#2237])
 
 ### Fixed
 - Seeking implementation in the stream cipher wrapper ([#2052])
 
 [#1759]: https://github.com/RustCrypto/traits/pull/1759
 [#2052]: https://github.com/RustCrypto/traits/pull/2052
+[#2237]: https://github.com/RustCrypto/traits/pull/2237
 
 ## 0.4.4 (2022-03-09)
 ### Changed